Use memset() to initialize statistics
This smell fishy, because it actually fixes a bug although the previous
initialization code should be fine.
Without memset(), the throughput of AES encryption on HiKey with ARM32
CE is unrealistically high (such as several thousands of MiB/s), the
stddev is big, and min is reported as 0 although update_stats() is
never called with 0.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>